requirements.txt
glfw==2.10.0
numpy==2.4.3
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
